new packages: hydra_imu (MinIMU9v3 drivers), hydra_common (utilities)
hydra_common currently includes only a version of the
`dynamic_reconfigure::Server<...>` template definition modified to play
nicely with nodelets.

hydra_imu contains both Arduino driver code and a small set of nodelets
to enable retrieving and processing data from the MinIMU9v3 units.
One of the salient features of this implementation is dynamic scale
selection in the Arduino code, which allows higher-resolution sensor
data when the measured quantity is small and higher-range when it's not.